feat: UseCaseTest 코드 작성

+++ apps/commerce-api/src/main/java/com/loopers/application/order/OrderCommand.java
@@ -1,10 +1,11 @@
 package com.loopers.application.order;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record OrderCommand(
         String userId,
         List<OrderItem> items
 ) {
-    public record OrderItem(Long productId, int quantity, int price) {}
+    public record OrderItem(Long productId, int quantity, BigDecimal price) {}
 }
