+++ apps/commerce-api/src/main/java/com/loopers/application/order/OrderInfo.java
@@ -1,12 +1,13 @@
 package com.loopers.application.order;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public record OrderInfo(
         Long orderId,
         String userId,
-        int totalAmount,
+        BigDecimal totalAmount,
         List<OrderItemInfo> items
 ) {
-    public record OrderItemInfo(Long productId, int quantity, int price) {}
+    public record OrderItemInfo(Long productId, int quantity, BigDecimal price) {}
 }
