+++ apps/commerce-api/src/main/java/com/loopers/interfaces/api/point/PointV1Controller.java
@@ -9,6 +9,8 @@ import com.loopers.support.error.ErrorType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
+
 @RequiredArgsConstructor
 @RestController
 @RequestMapping("/api/v1/points")
@@ -21,7 +23,7 @@ public class PointV1Controller implements PointV1ApiSpec {
     public ApiResponse<PointV1Dto.PointResponse> getPoints(
             @UserIdParam UserId userId
     ) {
-        Long points = pointUseCase.getPoints(userId);
+        BigDecimal points = pointUseCase.getPoints(userId);
         if (points == null) {
             throw new CoreException(ErrorType.NOT_FOUND, "존재하지 않는 ID입니다.");
         }
@@ -34,7 +36,7 @@ public class PointV1Controller implements PointV1ApiSpec {
             @UserIdParam UserId userId,
             @RequestBody PointV1Dto.PointChargeRequest request
     ) {
-        Long points = pointUseCase.chargePoints(userId, request.amount());
+        BigDecimal points = pointUseCase.chargePoints(userId, request.amount());
         return ApiResponse.success(PointV1Dto.PointResponse.from(points));
     }
 }
