+++ apps/commerce-api/src/main/java/com/loopers/interfaces/api/point/PointV1Dto.java
@@ -1,11 +1,13 @@
 package com.loopers.interfaces.api.point;
 
+import java.math.BigDecimal;
+
 public class PointV1Dto {
-    public record PointChargeRequest(Long amount) {
+    public record PointChargeRequest(BigDecimal amount) {
     }
 
-    public record PointResponse(Long pointValue) {
-        public static PointResponse from(Long pointValue) {
+    public record PointResponse(BigDecimal pointValue) {
+        public static PointResponse from(BigDecimal pointValue) {
             return new PointResponse(pointValue);
         }
     }
